[Multipaste] Record the time between pressing CTRL and V
UX/PM are researching an experimental key combo (hold ctrl for X time,
then press V to show multipaste).

Things UX/PM would like to know prior to creating this prototype:
1. Are users pressing Ctrl + V with generally consistent timing?
  - No current data here. This is important because we do not want to
    be annoying to users.
2. What time period would be reasonable for holding ctrl without
disrupting users?

After this metric lands, we will at least know the answer to #1, and may
be able to determine #2.

This key combo is just an idea as of now, but this data will lead to us
being able to build a prototype. There are a11y concerns with this
approach, but we likely will not get rid of the old key combo for
multipaste, but add this one as a faster way to show multipaste.

Bug: 1158619
Change-Id: Ie7b15b2330b1828e1fa99a46218514870efa74a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2616243
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844224}