Change AXLayoutObject::GetListStyle to work with custom counter styles
This patch stops the function from using the deprecated EListStyleType
enum, which only works with predefined counter styles. Now it's extended
to work with all counter styles that:
- For predefined counter styles, the return value remains the same
- For custom counter styles, the return value is kOther

Note: Ideally, AX should use the 'speak-as' descriptor to handle
custom counter styles. This is out of the scope of this patch.

Bug: 687225
Change-Id: I9b2065f9b025f268bbb9d44af9f7a1541ed19744
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2629991
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844229}