Include translate Extras in the Assistant voice search intent.
This includes the page URL and original/current/target language. It's gated by a new feature flag, AssistantIntentTranslateInfo, and is disabled by default.

I've opted to include the page URL when translate info is being added (rather than rely on the separate PAGE_URL Feature) because Assistant is required to return it in the Translate intent.

Bug: 1140581
Change-Id: Ic04777cb529902d4fdf461ca7e96c5b9dbcaaf9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2625961
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Josh Simmons <jds@google.com>
Cr-Commit-Position: refs/heads/master@{#844236}