WebUI NTP: load lazy bundle after timeout is reached
When creating many new tabs, requestIdleCallback() seems to hang until
there is a mouse action in the content area. This does not happen in the
normal use case of opening a single new tab and interacting with it.
Adding a timeout to requestIdleCallback() works around this issue.

Bug: 1167047
Change-Id: Ic10d7f75f0be0c6ce9b34992df1a4dc09b9d0e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2633068
Commit-Queue: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Auto-Submit: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844254}