webauthn: add UI for enterprise attestation.
With CTAP 2.1 it's possible for a security key to have an RP ID
hardcoded into it which can receive enterprise attestation on request.
If the RP ID is listed in the enterprise policy then this is permitted
automatically. Otherwise, we need to alert the user that this is
happening.

This change adds an attestation permission step to the WebAuthn UI which
acts the same as the current one, but with different wording to reflect
the greater impact of enterprise attestations.

Change-Id: I27fb9f7978b112910a51424f48c8e2567b146719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2616736
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844280}