[fido] Settings error when new PIN same as current
Show an error message on the Security Keys Set PIN settings page when
the new PIN is the same as the current PIN. Under some circumstances,
keys won't accept a PIN that is equal, so show an error early to be
safe.

The string provided was wider than the input, so to avoid shifting the
confirm input to the right, the error text is allowed to overflow with
white-space: nowrap (see the screenshot).

UXR review: go/force-pin-change-ux
Screenshot:
https://storage.cloud.google.com/chromium-translation-screenshots/a775ab5fe5a6302c4f83dc1b4863d21ff7b20ff0

Tbr: cpu@chromium.org
Fixed: 1152510
Change-Id: Id17f114c240469608fb0ef785bc5db4daba11a6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2615458
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844291}