[CrOS Cellular] Add flag to use external euicc
This CL adds a flag that when enabled will cause Cellular Setup
and Settings UI to use the external Euicc instead of the on-board
Euicc. This is useful for testing with test eSIM cards that can be
inserted in physical slots and appears as an additional Euicc in
Hermes.

Loadtime data value will be used in polymer elements in followup CL.

Bug: 1166697, 1093185
Change-Id: Iccccd2152f588dd7cae9f6d7fdd108795a46f118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2630431
Commit-Queue: Azeem Arshad <azeemarshad@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844292}