[ToolbarMVC] Move setUrl to LocationBarMediator
This move also allows/requires us to:
* Move backKeyPressed/mOriginalUrl/getOriginalUrl
* Stop overriding backKeyPressed and setUrl in SearchActivityLBL

In the latter case, we continue to override back key behavior via a
delegate. As far as I could tell it's safe to stop overriding setUrl,
since:
* SearchActivity's LocationBarDataProvider only ever returns an empty
UrlBarData
* SearchActivity's's UrlBar can't be unfocused
Just to be sure, I've added a test for this scenario.

Bug: 1147581
Change-Id: Ic1291b788a4913980d542135a828120e4a7cb353
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2594214
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844309}