Remove extra whitespace in DIR_METADATA files.
Changes were generated by a few lines of bash which I've documented at
http://go/paste/4886211966533632.

Limiting this CL to files in directories that don't set noparent in
OWNERS so it can be done with a small set of reviewers for a large set
of directories. Can do more targetted follow up CLs for the rest.

R=tedchoc

Bug: 1167291
Change-Id: I3ffbc694d284a34cbba1d14f36baecff0216da98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2633527
Commit-Queue: Andy Perelson <ajp@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844320}