[omnibox] [rich-autocompletion] Don't swap |fill_into_edit|.
Traditional AutcompleteMatch's set their |fill_into_edit| to their URLs.

With rich autocompletion before this CL, when the input matched the
title, |fill_into_edit| would instead be set to the title, and
|fill_into_edit_additional_text| would contain the URL and be
displayed in a views::Label adjacent to the omnibox. E.g., for the input
'ti', the location bar would display 'ti[tle] (url.com)'.

To avoid awkward cursor logic, only the omnibox is editable, and the
adjacent label containing |fill_into_edit_additional_text| is not. This
made it difficult to edit the suggestion URL before navigating to it
when the user input matched the suggestion title.

Therefore, this CL reverts to the original behavior of having
|fill_into_edit| be the URL. Specifically:
  1) No longer sets |fill_into_edit| to the suggestion title when the
input matches the title.
  2) Renames |fill_into_edit_additional_text| to |additional_text| since
it is no longer used when suggestion selection changes.
  3) Removes the |swapped_fill_into_edit| field as it would always be
false now.
  4) No longer sends the |additional_text| in view methods for selection
changes since we only want to display the additional text for the
default suggestions.
  5) Removes the |kRichAutocompletionShowTitlesParam| param as it would
require effort to make it compattible with the above changes, and
fishfooders concurred it's not promising.

This has the added side benefit of making |fill_into_edit| behavior
consistent between suggestions that seem the same. I.e., before when
selecting a suggestion whose title matched the input, sometimes the
location bar would display 'title (url.com)' and othertimes 'url.com'
(depending on the input length and whether the match was a shortcut
match).


Bug: 1163356, 1062446
Change-Id: If42f30d693b7d5204d94c3f21e72662879d8b857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2616973
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844333}