[fuchsia] Periodic memory pressure notifications
This makes Fuchsia behave more like Android memory pressure
signals. MODERATE and CRITICAL pressure notifications are now
sent periodically and not only when the level changes. This
is important as components expect to be able to just respond
to discrete events, which means we need periodic "refreshes"
of MODERATE and CRITICAL pressure.

Bug: 1166976
Change-Id: I829a3f3c6cd40e7753119ade19f2aac0d58a97ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2633682
Commit-Queue: David Reveman <reveman@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844371}