find-in-page: Fix Security DCHECK failure.
Fix Security DCHECK failure triggered by the conversion of RemoteFrame
to LocalFrame.

In OOPIFs mode, the text finder corresponding to the remote frame also
needs to be notified, the match rects are invalid and need to be
recalculated. Add a TODO first, will fix it in another CL.

Bug: 1167505,1147796
Change-Id: I7056ea45ef559988abc56fcbd1e4b02964ba29ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2631560
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: HuiWang <wanghui210@huawei.com>
Cr-Commit-Position: refs/heads/master@{#844375}