mac: Update hermetic Xcode to 12.3 and production SDK to 11.1.
Looks like response file support in ld64 might be less crashy in this
release, so we can try turning that back on after this roll.
(It's currently enabled for static libraries but not for shared
libraries or executables. We should first disable it for static
libraries to get a self-consistent state, and then try enabling it
for everything.)

Apparently `otool` now (sometimes?) calls `otool-classic`, so add
that to xcode_binaries.yaml and to content_shell_crash_test data.

Bug: 1147968
Change-Id: I90f45942e6f7d735b5fb3be242a517f002f2e072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2633748
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Auto-Submit: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844553}