[iOS Enterprise] Support dynamic refresh of BrowserSignin policy
Adds support for changing the value of the BrowserSignin policy to 0
while the browser is running.

Adds a BrowserAgent that monitors the
kSigninAllowed pref, which is controlled by the BrowserSignin policy.
When kSigninAllowed changes to false, the BrowserAgent dispatches a
command to SceneController to dismiss any in-progress sign-in UI and
sign the user out.

Bug: 1155745
Change-Id: Ifac4c785f73e098e94be8c3df35040ec60b489c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2616278
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#844583}