Switch to new virtuals for SkCanvas
One repercussion of this is that SkPaint no longer controls how images
are sampled -- that is done explicitly with a SkSamplingOptions
parameter to drawImage calls, and the same parameter passed to
image-shader factories.

To address this (ignoring paint's filter-quality), OpacityCanvas now
has to override more draw calls. This is fragile, and not the best
direction, but is necessary (it seems) at the moment.
See skbug.com/11126

Bug: skia:11126
Bug: skia:7650
Change-Id: I6562ca834ddde34c533e68fe4a5abb8c2fae22b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2605650
Commit-Queue: Mike Reed <reed@google.com>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844625}